# PERF: Improve RollingGroupby.count