--- a/pandas/core/window/common.py
+++ b/pandas/core/window/common.py
@@ -58,7 +58,6 @@ class WindowGroupByMixin(GotItemMixin):
         self._groupby.grouper.mutated = True
         super().__init__(obj, *args, **kwargs)
 
-    count = _dispatch("count")
     corr = _dispatch("corr", other=None, pairwise=None)
     cov = _dispatch("cov", other=None, pairwise=None)