- [x] closes #35625
- [x] tests added / passed
- [x] passes `black pandas`
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [x] whatsnew entry

```
In [1]: import pandas as pd
   ...:
   ...: # Generate sample df
   ...: df = pd.DataFrame({'column1': range(600), 'group': 5*['l'+str(i) for i in range(120)]})
   ...:
   ...: # sort by group for easy/efficient joining of new columns to df
   ...: df=df.sort_values('group',kind='mergesort').reset_index(drop=True)

In [2]: %timeit df['mean']=df.groupby('group').rolling(3,min_periods=1)['column1'].mean().values
5.59 ms ± 310 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [3]: %timeit df['sum']=df.groupby('group').rolling(3,min_periods=1)['column1'].sum().values
   ...:
5.34 ms ± 343 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [4]: %timeit df['count']=df.groupby('group').rolling(3,min_periods=1)['column1'].count().values
   ...:
4.97 ms ± 51.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```